split and update
