+++ Gopkg.toml
@@ -45,6 +45,12 @@ required = [
   # Needed because release 1.0.0 does not contain a LICENSE file
   revision = "4bbc89b6501cca7dd6b5557d78d70c8d2c6e8b97"
 
+[[override]]
+  name = "contrib.go.opencensus.io/exporter/stackdriver"
+  # HEAD as of 2019-02-11
+  # Needed because this includes a fix to support Stackdriver built-in metrics
+  revision = "c06c82c832edca4eaf7b0241bd655560a1be0346"
+
 [[constraint]]
   name = "github.com/knative/test-infra"
   branch = "master"
