+++ metrics/exporter.go
@@ -18,21 +18,15 @@ import (
 	"net/http"
 	"sync"
 
-	"contrib.go.opencensus.io/exporter/stackdriver"
-	"contrib.go.opencensus.io/exporter/stackdriver/monitoredresource"
-	"github.com/knative/pkg/metrics/metricskey"
-	"go.opencensus.io/exporter/prometheus"
 	"go.opencensus.io/stats/view"
-	"go.opencensus.io/tag"
 	"go.uber.org/zap"
 )
 
 var (
-	curMetricsExporter       view.Exporter
-	curMetricsConfig         *metricsConfig
-	curPromSrv               *http.Server
-	getMonitoredResourceFunc func(v *view.View, tags []tag.Tag) ([]tag.Tag, monitoredresource.Interface)
-	metricsMux               sync.Mutex
+	curMetricsExporter view.Exporter
+	curMetricsConfig   *metricsConfig
+	curPromSrv         *http.Server
+	metricsMux         sync.Mutex
 )
 
 // newMetricsExporter gets a metrics exporter based on the config.
@@ -62,138 +56,6 @@ func newMetricsExporter(config *metricsConfig, logger *zap.SugaredLogger) error 
 	setCurMetricsExporterAndConfig(e, config)
 	logger.Infof("Successfully updated the metrics exporter; old config: %v; new config %v", existingConfig, config)
 	return nil
-}
-
-func getKnativeRevisionMonitoredResource(gm *gcpMetadata) func(v *view.View, tags []tag.Tag) ([]tag.Tag, monitoredresource.Interface) {
-	return func(v *view.View, tags []tag.Tag) ([]tag.Tag, monitoredresource.Interface) {
-		tagsMap := getTagsMap(tags)
-		kr := &KnativeRevision{
-			// The first three resource labels are from metadata.
-			Project:     gm.project,
-			Location:    gm.location,
-			ClusterName: gm.cluster,
-			// The rest resource labels are from metrics labels.
-			NamespaceName:     valueOrUnknown(metricskey.LabelNamespaceName, tagsMap),
-			ServiceName:       valueOrUnknown(metricskey.LabelServiceName, tagsMap),
-			ConfigurationName: valueOrUnknown(metricskey.LabelConfigurationName, tagsMap),
-			RevisionName:      valueOrUnknown(metricskey.LabelRevisionName, tagsMap),
-		}
-
-		var newTags []tag.Tag
-		for _, t := range tags {
-			// Keep the metrics labels that are not resource labels
-			if !metricskey.KnativeRevisionLabels.Has(t.Key.Name()) {
-				newTags = append(newTags, t)
-			}
-		}
-
-		return newTags, kr
-	}
-}
-
-func getTagsMap(tags []tag.Tag) map[string]string {
-	tagsMap := map[string]string{}
-	for _, t := range tags {
-		tagsMap[t.Key.Name()] = t.Value
-	}
-	return tagsMap
-}
-
-func valueOrUnknown(key string, tagsMap map[string]string) string {
-	if value, ok := tagsMap[key]; ok {
-		return value
-	}
-	return metricskey.ValueUnknown
-}
-
-func getGlobalMonitoredResource() func(v *view.View, tags []tag.Tag) ([]tag.Tag, monitoredresource.Interface) {
-	return func(v *view.View, tags []tag.Tag) ([]tag.Tag, monitoredresource.Interface) {
-		return tags, &Global{}
-	}
-}
-
-func newStackdriverExporter(config *metricsConfig, logger *zap.SugaredLogger) (view.Exporter, error) {
-	setMonitoredResourceFunc(config, logger)
-	e, err := stackdriver.NewExporter(stackdriver.Options{
-		ProjectID:               config.stackdriverProjectID,
-		MetricPrefix:            config.domain + "/" + config.component,
-		GetMonitoredResource:    getMonitoredResourceFunc,
-		DefaultMonitoringLabels: &stackdriver.Labels{},
-	})
-	if err != nil {
-		logger.Error("Failed to create the Stackdriver exporter: ", zap.Error(err))
-		return nil, err
-	}
-	logger.Infof("Created Opencensus Stackdriver exporter with config %v", config)
-	return e, nil
-}
-
-func newPrometheusExporter(config *metricsConfig, logger *zap.SugaredLogger) (view.Exporter, error) {
-	e, err := prometheus.NewExporter(prometheus.Options{Namespace: config.component})
-	if err != nil {
-		logger.Error("Failed to create the Prometheus exporter.", zap.Error(err))
-		return nil, err
-	}
-	logger.Infof("Created Opencensus Prometheus exporter with config: %v. Start the server for Prometheus exporter.", config)
-	// Start the server for Prometheus scraping
-	go func() {
-		srv := startNewPromSrv(e)
-		srv.ListenAndServe()
-	}()
-	return e, nil
-}
-
-func getCurPromSrv() *http.Server {
-	metricsMux.Lock()
-	defer metricsMux.Unlock()
-	return curPromSrv
-}
-
-func resetCurPromSrv() {
-	metricsMux.Lock()
-	defer metricsMux.Unlock()
-	if curPromSrv != nil {
-		curPromSrv.Close()
-		curPromSrv = nil
-	}
-}
-
-func resetMonitoredResourceFunc() {
-	metricsMux.Lock()
-	defer metricsMux.Unlock()
-	if getMonitoredResourceFunc != nil {
-		getMonitoredResourceFunc = nil
-	}
-}
-
-func setMonitoredResourceFunc(config *metricsConfig, logger *zap.SugaredLogger) {
-	metricsMux.Lock()
-	defer metricsMux.Unlock()
-	if getMonitoredResourceFunc == nil {
-		gm := retrieveGCPMetadata()
-		metricsPrefix := config.domain + "/" + config.component
-		logger.Infof("metrics prefix: %s", metricsPrefix)
-		if metricskey.KnativeRevisionMetricsPrefixes.Has(metricsPrefix) {
-			getMonitoredResourceFunc = getKnativeRevisionMonitoredResource(gm)
-		} else {
-			getMonitoredResourceFunc = getGlobalMonitoredResource()
-		}
-	}
-}
-
-func startNewPromSrv(e *prometheus.Exporter) *http.Server {
-	sm := http.NewServeMux()
-	sm.Handle("/metrics", e)
-	metricsMux.Lock()
-	defer metricsMux.Unlock()
-	if curPromSrv != nil {
-		curPromSrv.Close()
-	}
-	curPromSrv = &http.Server{
-		Addr:    ":9090",
-		Handler: sm,
-	}
-	return curPromSrv
 }
 
 func getCurMetricsExporter() view.Exporter {
