+++ metrics/prometheus_exporter.go
@@ -1,0 +1,67 @@
+/*
+Copyright 2019 The Knative Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package metrics
+
+import (
+	"net/http"
+
+	"go.opencensus.io/exporter/prometheus"
+	"go.opencensus.io/stats/view"
+	"go.uber.org/zap"
+)
+
+func newPrometheusExporter(config *metricsConfig, logger *zap.SugaredLogger) (view.Exporter, error) {
+	e, err := prometheus.NewExporter(prometheus.Options{Namespace: config.component})
+	if err != nil {
+		logger.Error("Failed to create the Prometheus exporter.", zap.Error(err))
+		return nil, err
+	}
+	logger.Infof("Created Opencensus Prometheus exporter with config: %v. Start the server for Prometheus exporter.", config)
+	// Start the server for Prometheus scraping
+	go func() {
+		srv := startNewPromSrv(e)
+		srv.ListenAndServe()
+	}()
+	return e, nil
+}
+
+func getCurPromSrv() *http.Server {
+	metricsMux.Lock()
+	defer metricsMux.Unlock()
+	return curPromSrv
+}
+
+func resetCurPromSrv() {
+	metricsMux.Lock()
+	defer metricsMux.Unlock()
+	if curPromSrv != nil {
+		curPromSrv.Close()
+		curPromSrv = nil
+	}
+}
+
+func startNewPromSrv(e *prometheus.Exporter) *http.Server {
+	sm := http.NewServeMux()
+	sm.Handle("/metrics", e)
+	metricsMux.Lock()
+	defer metricsMux.Unlock()
+	if curPromSrv != nil {
+		curPromSrv.Close()
+	}
+	curPromSrv = &http.Server{
+		Addr:    ":9090",
+		Handler: sm,
+	}
+	return curPromSrv
+}
